scraping/requirements.txt
urlopen
beautifulsoup4
psycopg2-binary
pandas
dotenv
chardet